refactor: 마이페이지 리팩토링
