+++ README.md
@@ -72,16 +72,18 @@
 ### 1. 개발당사자에 따라 무한스크롤이 작동하지 않는 문제
 
 ![트러블슈팅1_1](https://github.com/user-attachments/assets/814b88c6-18e9-4e12-8563-96ee56482d6c)
-이유모를 json-server의 버전 차이로 인해 useInfiniteQuery의 return값에 차이가 발생하여 page관리의 방법이 달라야 했습니다.
+json-server의 버전 차이로 인해 useInfiniteQuery의 return값에 차이가 발생하여 page관리의 방법이 달라야 했습니다.
 ![트러블슈팅1_2](https://github.com/user-attachments/assets/af1ef95d-7bab-4688-b58f-516c0231ce9c)
 무한스크롤 데이터 요청의 page관리 방법을 변경하여 정상 작동하도록 수정하였습니다.
 ![트러블슈팅1_3](https://github.com/user-attachments/assets/3d8fd192-782c-4014-b9f8-31d0d73329bd)
 
 <br />
 
-### 2.
+### 2. 지도 생성 함수를 한 페이지의 두 컴포넌트에서 호출 시 지도가 불러와지지 않는 문제
 
-=> <br/>
+![트러블슈팅2_1](https://github.com/user-attachments/assets/8dfbac11-fc29-4602-a072-8208ece2b341)
+함수 내부에서 getElementById로 id값을 사용하는데, 두 컴포넌트가 같은 id를 지니고 있어 앞쪽 요소만 지도가 담겼습니다.
+id생성방식을 변경하여 다른 id값을 지니게 해 문제를 해결하였습니다.
 
 <br />
 
@@ -118,7 +120,7 @@
 ### 전상국
 
 ```
-
+ 프로젝트에 늦게 참여하여 SA를 참여하지 못했고 그에따라 프로젝트 파악이 어려웠는데, 팀원분들의 배려와 도움으로 빠르게 적응할 수 있었으며, 소통을 통해 여러 문제를 해결할 수 있어 좋았습니다.
 ```
 
 <br/>
@@ -134,7 +136,7 @@
 ### 설하영
 
 ```
-
+ 지난 뉴스피드 과제에서는 어떤형태의 페이지를 만드는지 방향성이 있었다면, 이번 아웃소싱 과제에서는 api를 선택하여 어떤 페이지를 만들어야 하는지부터 기획을 했어야 해서 조금 막막했었습니다. 그치만 팀원분들과 이런저런 아이디어를 내면서 기획을 하고 막히는 부분이 있을때 마다 소통하면서 문제를 해결해나갈 수 있어서 힘들었지만 재미있었습니다.
 ```
 
 <br/>
