+++ src/queries/query.keys.js
@@ -1,0 +1,5 @@
+export const queryKeys = {
+  boardController: {
+    myFeeds: () => ["myFeeds"]
+  }
+};
